% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark-package.R
\name{cmark_version}
\alias{cmark_version}
\title{Version of the Underlying cmark Library}
\usage{
cmark_version()
}
\value{
A \link[base:numeric_version]{numeric_version} object.
}
\description{
Version of the Underlying cmark Library
}
