% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mark.R
\name{coef.mark}
\alias{coef.mark}
\title{MARK model beta parameters}
\usage{
\method{coef}{mark}(object,...)
}
\arguments{
\item{object}{a MARK model object}

\item{...}{additional non-specified argument for S3 generic function}
}
\value{
A vector or dataframe of beta estimates
}
\description{
MARK model beta parameters
}
\author{
Jeff Laake
}
\keyword{utility}
