% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.mark.mcmc.R
\name{create.mark.mcmc}
\alias{create.mark.mcmc}
\title{Create mcmc object for analysis with coda}
\usage{
create.mark.mcmc(filename, ncovs, nmeans, ndesigns, nsigmas, nrhos, nlogit,
  include = F)
}
\arguments{
\item{filename}{name of file containing mcmc values}

\item{ncovs}{number of covariates}

\item{nmeans}{number of means}

\item{ndesigns}{number of designs}

\item{nsigmas}{number of sigmas}

\item{nrhos}{number of rhos}

\item{nlogit}{number of logits}

\item{include}{if TRUE it includes ir/propJumps fields}
}
\value{
An mcmc object if one chain and an mcmc list if more than one chain. Each can
be used with the coda package
}
\description{
Reads in mcmc file from program MARK in binary and returns an mcmc object that can be used with coda functions which are 
most easily accessed via codamenu().
}
\author{
Jeff Laake
}
\keyword{utility}
