\name{load.model}
\alias{load.model}
\title{Load external model}
\description{
Loads external model into workspace with name model}
\usage{
load.model(model)
}
\arguments{
  \item{model}{name of MARK model object stored externally}
 }
\value{
None; side effect only to load object with name model (not name given to model)
}
\author{Jeff Laake}
\keyword{utility}
