\name{logitCI}
\alias{logitCI}
\title{Logit Confidence Interval}
\description{
Computes confidence interval of real parameter (bounded between 0-1) using logit transform
}
\usage{
logitCI(x,se)
}
\arguments{
  \item{x}{vector of real estimates}
  \item{se}{vector of se estimates}
}
\value{A dataframe with x,se and conf interval
}
\author{Jeff Laake}
\keyword{utility}
