% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark.R
\name{md_is}
\alias{md_is}
\alias{md_is_block}
\alias{md_is_inline}
\alias{md_is_leaf}
\title{Classification}
\usage{
md_is_block(x)

md_is_inline(x)

md_is_leaf(x)
}
\arguments{
\item{x}{A markdown node.}
}
\description{
Classification
}
