% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark.R
\name{md_iterate}
\alias{md_iterate}
\title{Iteration}
\usage{
md_iterate(x, callback)
}
\arguments{
\item{x}{A markdown node.}

\item{callback}{A function.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Iteration
}
\examples{
root <- parse_md("# Hello")
md_iterate(root, function(node, event) {
  cat(event, "ing ", format(node), sep = "")
})
}
