% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark.R
\name{md_location}
\alias{md_location}
\alias{md_start_line}
\alias{md_start_column}
\alias{md_end_line}
\alias{md_end_column}
\title{Location in Source Document}
\usage{
md_start_line(x)

md_start_column(x)

md_end_line(x)

md_end_column(x)
}
\arguments{
\item{x}{A markdown node.}
}
\value{
An integer.
}
\description{
Location in Source Document
}
