% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark.R
\name{md_node}
\alias{md_node}
\alias{md_type}
\alias{md_first_child}
\alias{md_literal}
\title{Markdown nodes}
\usage{
md_type(x)

md_first_child(x)

md_literal(x)
}
\arguments{
\item{x}{A markdown node.}
}
\description{
Markdown nodes
}
