% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark.R
\name{md_traversal}
\alias{md_traversal}
\alias{md_next}
\alias{md_previous}
\alias{md_parent}
\alias{md_first_child}
\alias{md_last_child}
\title{Tree Traversal}
\usage{
md_next(x)

md_previous(x)

md_parent(x)

md_first_child(x)

md_last_child(x)
}
\arguments{
\item{x}{A markdown node.}
}
\description{
Tree Traversal
}
