% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.average.R
\name{model.average}
\alias{model.average}
\alias{model.average.default}
\title{Compute model averaged estimates}
\usage{
model.average(x, ...)
}
\arguments{
\item{x}{is either a list with a prescribed structure as defined in
\code{\link{model.average.list}} or a \code{marklist} as described in
\code{\link{model.average.marklist}}}

\item{...}{additional arguments passed to specific functions}
}
\value{
The structure of the returned value depends on which function is
called.
}
\description{
A generic function to compute model averaged estimates and their standard
errors or variance-covariance matrix.
}
\seealso{
\code{\link{model.average.marklist}},\code{\link{model.average.list}}
}
\author{
Jeff Laake
}
\keyword{utility}
