% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMark-package.R
\docType{data}
\name{NSpeciesOcc}
\alias{NSpeciesOcc}
\title{Multiple Species Occupancy}
\format{A data frame with 32 observations of 2 variables 
 \describe{ 
 \item{ch}{a character vector containing the capture history (each is 2 character positions) for 4 occasions with 5 species}
 \item{freq}{capture history frequency} 
 }}
\description{
Example expected value data generated by Gary White for testing.
}
\examples{
\donttest{
# read in expected value data from Gary White
data("NSpeciesOcc")
# process data and specify number of species=5 in the mixtures argument
dp=process.data(NSpeciesOcc,model="NSpeciesOcc",mixtures=5)
# make design data and fit model used to generate the data
ddl=make.design.data(dp)
model=mark(dp,ddl,model.parameters=list(f=list(formula=~-1+mixture,link="sin"),
p=list(formula=~1,link="sin")))
}
}
\author{
Jeff Laake
}
\keyword{datasets}
