% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark.R
\name{parse_md}
\alias{parse_md}
\title{Parse Markdown text}
\usage{
parse_md(x)
}
\arguments{
\item{x}{A character vector of lines of Markdown.}
}
\value{
A markdown node.
}
\description{
Parse Markdown text
}
\examples{
parse_md("# Hello")
}
