% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.mark.R
\name{print.summary.mark}
\alias{print.summary.mark}
\title{Prints summary of MARK model parameters and results}
\usage{
\method{print}{summary.mark}(x,...)
}
\arguments{
\item{x}{list resulting from call to \code{summary}}

\item{...}{additional non-specified argument for S3 generic function}
}
\value{
None
}
\description{
Prints summary of MARK model parameters and results
}
\seealso{
summary.mark
}
\author{
Jeff Laake
}
\keyword{utility}
