% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark.R
\name{read_md}
\alias{read_md}
\title{Read a Markdown file}
\usage{
read_md(x)
}
\arguments{
\item{x}{A file path or \link[base:connections]{connection}.}
}
\value{
A markdown node.
}
\description{
Read a Markdown file
}
