% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark.R
\name{render_md}
\alias{render_md}
\title{Render Markdown}
\usage{
render_md(x, width = getOption("width"))
}
\arguments{
\item{x}{A markdown node.}

\item{width}{An integer specifying the maximum line width of the output text.}
}
\description{
Render Markdown
}
