% Generated by roxygen2 (4.0.1): do not edit by hand
\name{valid.parameters}
\alias{valid.parameters}
\title{Determine validity of parameters for a model (internal use)}
\usage{
valid.parameters(model, parameters)
}
\arguments{
\item{model}{type of c-r model ("CJS", "Burnham" etc)}

\item{parameters}{vector of parameter names (for example "Phi" or "p" or
"S")}
}
\value{
Logical; TRUE if all parameters are acceptable and FALSE otherwise
}
\description{
Checks to make sure specified parameters are valid for a particular type of
model.
}
\author{
Jeff Laake
}
\seealso{
\code{\link{setup.parameters}}, \code{\link{setup.model}}
}
\keyword{utility}

