% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.model}
\alias{load.model}
\title{Load external model}
\usage{
load.model(model)
}
\arguments{
\item{model}{name of MARK model object stored externally}
}
\value{
None; side effect only to load object with name model (not name
given to model)
}
\description{
Loads external model into workspace with name model
}
\author{
Jeff Laake
}
\keyword{utility}

