% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RMark-package.R
\docType{data}
\name{LogitNormalMR}
\alias{LogitNormalMR}
\title{Example of LogitNormal Mark-Resight model}
\format{A data frame with 35 observations on the following variable.
\describe{ \item{ch}{a character vector} }}
\description{
Data and example illustrating LogitNormal Mark-Resight model.
}
\examples{
data(LogitNormalMR)
logitNor.proc=process.data(LogitNormalMR,model="LogitNormalMR",
		counts=list("Unmarked seen"=c(96,68,59),
				    "Marked Unidentified"=c(0,0,0,0,1,1,1,0,0,3,0,1)),
			         time.intervals=c(0,0,0,1,0,0,0,1,0,0,0))
logitNor.ddl=make.design.data(logitNor.proc)
mod=mark(logitNor.proc,logitNor.ddl,threads=2)
summary(mod)
}
\keyword{datasets}

