% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark.R
\name{md_node}
\alias{md_node}
\alias{md_type}
\alias{md_literal}
\alias{md_literal<-}
\alias{md_heading_level}
\alias{md_heading_level<-}
\alias{md_list_type}
\alias{md_list_type<-}
\alias{md_list_delim}
\alias{md_list_delim<-}
\alias{md_list_start}
\alias{md_list_start<-}
\alias{md_list_tight}
\alias{md_list_tight<-}
\alias{md_fence_info}
\alias{md_fence_info<-}
\alias{md_url}
\alias{md_url<-}
\alias{md_title}
\alias{md_title<-}
\title{Accessors}
\usage{
md_type(x)

md_literal(x)

md_literal(x) <- value

md_heading_level(x)

md_heading_level(x) <- value

md_list_type(x)

md_list_type(x) <- value

md_list_delim(x)

md_list_delim(x) <- value

md_list_start(x)

md_list_start(x) <- value

md_list_tight(x)

md_list_tight(x) <- value

md_fence_info(x)

md_fence_info(x) <- value

md_url(x)

md_url(x) <- value

md_title(x)

md_title(x) <- value
}
\arguments{
\item{x}{A markdown node.}

\item{value}{A string.}
}
\description{
Accessors
}
