% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark.R
\name{md_node}
\alias{md_node}
\alias{md_type}
\alias{md_literal}
\alias{md_literal<-}
\alias{md_heading_level}
\alias{md_heading_level<-}
\title{Accessors}
\usage{
md_type(x)

md_literal(x)

md_literal(x) <- value

md_heading_level(x)

md_heading_level(x) <- value
}
\arguments{
\item{x}{A markdown node.}

\item{value}{A string.}
}
\description{
Accessors
}
