% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.ch.R
\name{process.ch}
\alias{process.ch}
\title{Process release-recapture history data}
\usage{
process.ch(ch, freq = NULL, all = FALSE)
}
\arguments{
\item{ch}{Vector of character strings; each character string is 
composed of either a constant length sequence of single characters (01001) or 
the character string can be comma separated if more than a single 
character is used (1S,1W,0,2W). If comma separated, each string must contain a constant
number of elements.}

\item{freq}{Optional vector of frequencies for ch; if missing assumed to be
a; if <0 indicates a loss on capture}

\item{all}{FALSE is okay for cjs unless R code used to compute lnl instead of
FORTRAN; must be true for js because it returns additional quantities needed for entry prob.}
}
\value{
\item{nocc}{number of capture occasions} \item{freq}{absolute value
of frequency for each ch} \item{first}{vector of occasion numbers for first
1} \item{last}{vector of occasion numbers for last 1} \item{loc}{vector of
indicators of a loss on capture if set to 1} \item{chmat}{capture history
matrix} \item{FtoL}{1's from first (1) to last (1) and 0's elsewhere; only
if all==TRUE} \item{Fplus}{1's from occasion after first (1) to nocc(last
occasion); only if all==TRUE} \item{Lplus}{1's from occasion after last (1)
to nocc; only if all==TRUE} \item{L}{1's from last (1) to nocc; only if
all==TRUE}
}
\description{
Creates needed constructs from the release-recapture history.
}
\author{
Jeff Laake
}
