% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mark.binary.R
\name{read.mark.binary}
\alias{read.mark.binary}
\alias{read.mark.binary.linux}
\title{Reads binary file output from MARK and returns a list of the results}
\usage{
read.mark.binary(filespec, derived_labels)
}
\arguments{
\item{filespec}{Filename specification for binary output file from
MARK;named here as markxxx.vcv}

\item{derived_labels}{vector of labels for derived parameters; NULL if no derived parameters for model}
}
\value{
List of estimates, se, lcl, ucl and var-cov matrices for beta, real
and derived estimates \item{beta}{Dataframe for beta parameters containing
estimates, se, lcl, ucl} \item{beta.vcv}{variance-covariance matrix for beta
estimates} \item{real}{Dataframe for real parameters containing estimates,
se, lcl, ucl} \item{real.vcv}{variance-covariance matrix for real estimates}
\item{derived}{Dataframe for derived parameters (if any) containing
estimates, se, lcl, ucl} \item{derived.vcv}{variance-covariance matrix for
derived estimates (if any)}
}
\description{
Window and linux versions to read binary files created by MARK
}
\seealso{
\code{\link{extract.mark.output}}
}
\author{
Jeff Laake
}
\keyword{utility}
