% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark.R
\name{render_md}
\alias{render_md}
\title{Render Markdown}
\usage{
render_md(x, ..., format = "commonmark", width = getOption("width"))
}
\arguments{
\item{x}{A markdown node.}

\item{...}{Arguments reserved for future use.}

\item{format}{A string.}

\item{width}{An integer.}
}
\description{
Render Markdown
}
