% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.r
\name{store}
\alias{store}
\alias{restore}
\title{Store models externally or restore to workspace from external storage}
\usage{
store(x)
}
\arguments{
\item{x}{marklist of models}
}
\value{
A modified marklist to replace the previous marklist specified as
the argument.
}
\description{
Stores/restores all mark model objects in a marklist either to or from
external storage.
}
\details{
For \code{store}, each mark model is stored externally and the object in the
list is replaced with the filename of the object. \code{restore} does the
opposite of storing the saved external object into the marklist and then
deleting the saved file.
}
\author{
Jeff Laake
}
\keyword{utility}
